fix: make sure to logout completely
previously the logout-button only logged the user out for next-auth, not Keycloak. When logging in and logging out the user therefore could not choose to log in with a different account. The user also did not have to type in any credentials. To fix this, an event handler was added to the next-auth signOut-event that sends a logout-request to the keycloak-instance